rename rokit_locator_api.py to talk_to_rokit_locator.py
