+++ test_api.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from rokit_locator_api import *
+from talk_to_rokit_locator import *
 
 # get_client_localization_pose("localhost", 9011)
 get_client_control_mode("localhost", 9004)
